rewrite everything for now 2.0
