+++ api/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "trends",
+  "scripts": {
+    "dev": "node ./index.js",
+    "start": "NODE_ENV=production node ./index.js"
+  },
+  "dependencies": {
+    "apollo-boost": "~0.1.9",
+    "apollo-server-express": "~1.3.6",
+    "await-wrap": "~0.1.0",
+    "body-parser": "~1.18.3",
+    "cors": "~2.8.4",
+    "express": "~4.16.3",
+    "graphql": "~0.13.2",
+    "graphql-tag": "~2.9.2",
+    "graphql-tools": "~3.0.2",
+    "isomorphic-unfetch": "~2.0.0",
+    "lru-cache": "~4.1.3",
+    "querystring": "~0.2.0"
+  },
+  "devDependencies": {
+    "babel-plugin-emotion": "~9.0.1",
+    "babel-plugin-transform-flow-strip-types": "6.22.0",
+    "flow-bin": "~0.74.0",
+    "husky": "~0.14.3",
+    "prettier": "~1.13.7",
+    "pretty-quick": "~1.6.0"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "pretty-quick --staged"
+    }
+  }
+}
