+++ package.json
@@ -3,36 +3,9 @@
   "scripts": {
     "dev": "node ./server/index.js",
     "build": "next build",
-    "start": "NODE_ENV=production node ./server/index.js",
-    "flow": "flow",
-    "test": "npm run flow"
-  },
-  "dependencies": {
-    "apollo-boost": "~0.1.9",
-    "apollo-server-express": "~1.3.6",
-    "await-wrap": "~0.1.0",
-    "body-parser": "~1.18.3",
-    "cors": "~2.8.4",
-    "emotion": "~9.0.2",
-    "emotion-server": "~9.0.2",
-    "express": "~4.16.3",
-    "global": "~4.3.2",
-    "graphql": "~0.13.2",
-    "graphql-tag": "~2.9.2",
-    "graphql-tools": "~3.0.2",
-    "isomorphic-unfetch": "~2.0.0",
-    "lru-cache": "~4.1.3",
-    "next": "^7.0.2",
-    "next-offline": "^3.2.1",
-    "querystring": "~0.2.0",
-    "react": "^16.6.3",
-    "react-apollo": "~2.1.5",
-    "react-dom": "^16.6.3",
-    "react-emotion": "~9.0.2"
+    "start": "NODE_ENV=production node ./server/index.js"
   },
   "devDependencies": {
-    "babel-plugin-emotion": "~9.0.1",
-    "babel-plugin-transform-flow-strip-types": "6.22.0",
     "flow-bin": "~0.74.0",
     "husky": "~0.14.3",
     "prettier": "~1.13.7",
@@ -41,14 +14,6 @@
   "husky": {
     "hooks": {
       "pre-commit": "pretty-quick --staged"
-    }
-  },
-  "now": {
-    "alias": [
-      "trends.now.sh"
-    ],
-    "env": {
-      "GITHUB_ACCESS_TOKEN": "@github-access-token"
     }
   }
 }
