+++ www/next.config.js
@@ -3,6 +3,8 @@ const withOffline = require('next-offline')
 
 const isDev = process.env.NODE_ENV !== 'production'
 
+console.log(process.env.NOW_URL)
+
 module.exports = withOffline({
   publicRuntimeConfig: {
     api: isDev ? 'http://localhost:3000' : typeof window !== 'undefined' ? '' : process.env.NOW_URL,
