+++ www/static/apple-touch-icon-114x114.png
