+++ www/static/apple-touch-icon-152x152.png
