+++ www/static/favicon-96x96.png
