+++ www/static/favicon.ico
