+++ www/static/manifest.json
